move duration to sqlutil
