+++ pkg/sqlutil/duration_test.go
@@ -1,4 +1,4 @@
-package models
+package sqlutil
 
 import (
 	"encoding/json"
