+++ pkg/utils/utils.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"math"
 	mrand "math/rand"
-	"strings"
 	"time"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/services"
@@ -49,12 +48,4 @@ func ContextWithDeadlineFn(ctx context.Context, deadlineFn func(orig time.Time) 
 func IsZero[C comparable](val C) bool {
 	var zero C
 	return zero == val
-}
-
-// EnsureHexPrefix adds the prefix (0x) to a given hex string.
-func EnsureHexPrefix(str string) string {
-	if !strings.HasPrefix(str, "0x") {
-		str = "0x" + str
-	}
-	return str
 }
